conductor/requirements.txt
transformers==4.52.4
datasets==3.6.0
accelerate
bitsandbytes
peft
hf-transfer==0.1.9
deepspeed
trl==0.19.1
hydra-core==1.3.2
ray[serve]
matplotlib
wandb
tensorboard==2.18.0
scikit-learn
evaluate
tenacity==9.1.2
google-genai
openai
anthropic
word2number
boto3==1.40.4
tenacity==8.5.0